Change of semantic for path type variables to better accomodate links
with this commit, the type `path` still has two parameters:
- first is named "kind" and says whether the path is absolute or
  relative;
- second is named "object" and describes the file system object
  identified by the path (file or dir)

"kind" as defined here is a syntactic characteristic of the path,
taken as an identifier, while "object" is a semantic notion
(i.e. related to the interpretation of the path). From that remark, I
modified the behaviour of `Link` items:
- `Link` items are relative items whatever the path they target
  to. This is because the kind is useful to know if a path can be
  interpreted or not. A single `Link` item cannot so it should be
  flagged relative
- a link, viewed as an object in a file system inherits the properties
  from the object they point to. For instance a link pointing to a
  directory should be usable as a left argument for concat. So a link
  should have the same "object" type than the path it points to.

This typing uses only 2 variables (like the initial one) but is better
shaped IMO to express the constraints related to concat, and the
constraint that if a path is flagged "abs" then it can be interpreted
in a file system.

The only drawback I see is the type of resolve: it is not possible
anymore to derive the type of its return value from the argument. So
we have to use another GADT to force a user to consider all cases.