pure/Path: introduced a function to map path whose kind is unknown
in the case, the kind of a path is not statically known, but doesn't
matter to compute a result, we have to pattern match on the path
anyway and duplicate the code in the two branches (Abs_path and
Rel_path). To avoid that, I introduced a map function for values of
type `of_some_kind`. The mapped function should be passed as a record
(the type is called `mapper`) to enforce universally-quantified
polymorphism.

The result is rather concise (better than pattern matching for sure)
and does avoid code duplication.