unix/Filesys: partial implementation of mkdir
This is a first implementation of `mkdir`, a function that creates a
directory specified with a path. This function recursively creates
directories and links in order to create the desired path.

The function is provided with a test that randomly generates a
thousand abs dir paths, creates them using `mkdir` and tests they are
correctly created using `exists`.

BTW I fixed a bug related to the use of `Unix.stat` instead of
`Unix.lstat`. Using the former, links were not recognized as such and
since we provide a normalize function, it makes more sense to provide
only `lstat`.

Note that the function is not finished yet, as two cases are not
properly handled yet:
  - cycles: `mkdir (Cons (Dir "foo", Item (Link (bar, Item Dotdot))))`
    does not terminate.
  - self-contradicting paths:
    Cons (Dir "foo", Cons (Dotdot, Item (File "foo"))) would return a
    Unix error.

Not sure if we want to detect the second case though.