port tests to async
Define (>::=) operator, which allows OUnit testing of Async
computations. Convert all code as necessary to make it compile against
the library, which was ported to Async in
dc22fa65d5f9ddb7943e9af7f75ea5141414592f.

Tests are not passing, but not sure if it is due to an error in this
commit. Checking in this code anyway, so we can incrementally fix.