tests for path deserialization functions + fixed 2 bugs (#8)
These are tests for [Path.(rel_){dir,file}_path_of_sexp]. I added
random generation function for file paths and a simple test that
serializes then deserializes absolute paths and checks that they are
equal.

I thus found two bugs in the deserialization functions which are fixed
in this commit.