Introduction of broken links in path representation
[Phat_path.t] now has a case for broken links, that is links whose
target should not exist. This was created because it is sometimes
useful to represent the path of a link, even if it cannot be resolved
First encountered use case is the fold function: you want to iterate
on a link even if it's broken, but for that you need to be able to
construct its [Phat_path.t] representation.

In addition a new [Pure_phat.link] phantom type was introduced, in the
same line than [Pure_phat.file] and [Pure_phat.dir]. Remind that the
['typ] parameter of [Pure_phat.t] represents the type of object
identified by a path after resolution. The ['typ] of broken link is
[link]: they don't resolve to any defined target so they
conventionnaly resolve to them-selves, like regular files/dir.