An implementation of resolve immune to cyclic values
The previous implementatioin of `resolve` and `resolve_any` does not
terminate on paths that are built using mutually recursive
values. This implementation detects cyclicity using a set that keeps
track of paths that we started to resolve but did not yet
finished. That is the set of paths we want to resolve in the end. Thus
if we start to say that we need a path `p` to resolve a set of paths
containing `p`, then it means there's a cycle and that the link is
broken.

So far so good. The trouble now is that resolution may change the
[typ] of a path: you have a link eventually representing a dir : `p :
(abs, dir) path` which happens to be cyclic (and thus broken). After
resolution you end with `q : (abs, link) path`. This is a bit
unfortunate because resolution doesn't maintain kind nor type of a
path, making its use less confortable.

Tests don't compile any more, I want to try another approach before
fixing them.