A new approach to handle cycles in path: smart constructor for links
The previous commit introduced a resolution function that would cope
for paths made of cyclic ocaml values. Not only is this not so easy to
implement, but the same thing has to be done for other operations,
like `typ_of` for instance, but it also doesn't protect a user when
she wants to recursively traverse a path. She would have to implement
the same kind of protection. This seems like a lot of hassle for a
pretty minor feature (since many operations are not defined on cyclic
paths anyway).

In this commit, I introduce an alternative approach that consists in
setting the item type private and defining (smart) constructors for
it. The only difficult case is for links, where the constructor
detects cyclic values and returns a broken link to the user.

Those two measures are the basis of the following invariant: paths
(seen as ocaml values) cannot be cyclic. The immediate consequence is
that one can traverse them without worries.

The implementation of `resolve` is back to the way it was, and what's
even better to its original type (since the case of broken links is
handled at link creation).