Remove protection against cyclic path expressions
In 89fd620 I introduced smart constructors to make the construction of
cyclic ocaml values of type Path.t impossible. This seems overkill, as
the construction of such values cannot happen "just by mistake", and
has little practical interest. Other libraries (even List in std) do
not bother with this and we should do the same.

So in this commit, I basically undid 89fd620.