async-unix/Filesys.exists: removed useless protections against cyclic values
As decided in 267b347, cyclic ocaml values for paths are now
unsupported and not checked. The implementation of `Filesys.exists`
was then unnecessarily complicated. Since we basically traverse a
`Path.t` value, which is assumed to be a tree or at worse a DAG, the
traversal is necessarily finite.

Only (small) drawback is that for DAGs we perform useless tests calls
to `Sys.file_exists`. I feel they are really not worth a complicated
implementation.