tests: fixed random path generation
We need to generate paths that correspond to objects under a certain
directory (typically /tmp/ounitXXX). In particular we must be careful
not to add too many "..", otherwise the generated path could point
outside the test directory. This is made a bit tricky because of
links, and the previous version would not work very well in that
respect.

Previously I limited the complexity of generated paths for mkdir and
exists tests to avoid the problem, but this commit lifts that
limitation. We now generate a wider variety of paths, such that all
objects mentionned in the paths lie in the test directory.