Filesys.reify: bug fix when reifying links
In `reify_link`, I forgot to use `Path.make_relative` to make sure
that the relative target of a path is reified into a relative path
also.

This fixes the bug found in the previous commit, but shows a new one:
`make_relative` is not smart enough for reify to work correctly and
produces unnecessarily long relative paths.