Path.make_relative: simpler and smarter
The function for detecting a common prefix between two paths can now
detect cases with all kinds of items instead of Dir only.

This fixes the problem described in the previous commit but... has
shown a new one, which is that the `reify` follows links by
concatenating the dirname of a link and its target (correct) and
normalize the result (incorrect, because normalize does not guarantee
that the resulting path still identifies the same object on the file
system in presence of links).