Filesys.reify: fixed incorrect way of following links
I would use normalize in `reify` to avoid dealing with Dot and
Dotdot. In particular I would follow (rel) links by concatenating with
the dirname of the link and its target (correct) then normalize
(incorrect, at least till #18 is not addressed).

This commit removes uses of `normalize` and deals with Dot and Dotdot
cases (which were not difficult at all, in the end).