tests: introduced random generation of normalized path (and fixed reify test)
The test for reify works as follows:
  - generate a random path p
  - mkdir it
  - parse its string representation as q
  - reify q as r
  - check that r exists exactly on the file system

This test would fail for no good reason, namely that reify returns a
normalized path, although p wasn't. This showed up in link target
(which could be non normal) and exists would return
`Yes_as_other_object since the target were not identical.

In order to side step the problem, I introduced functions that
generate normalized path to begin with, and this time the test passes.

Oh, and I've looked at the generated examples: they are a plain
nightmare :o)