async/Filesys.fold: faithful representation of links
up to now, fold would build links it encountered in a simplistic way:
use the target (obtained in a string form) to find if the link
represents a dir or a file, then just build the link by parsing the
target without checking if the target itself contains links. The
problem with this is that we pass a representation of the link which
is not accurate, although it is equivalent modulo resolution.

Thanks to the new reify function, the target is itself (recursively)
analysed to build it exactly as it is on the filesystem.