Reformulated test for fold
The former version would work on sexp representations, now works on
GADT values directly, which seems cleaner to me.

BTW the expected behaviour in the test was wrong (too many initial
dots to the relative paths), and so it the behaviour of fold. The test
has been fixed but not the function yet, so the test fails for now.