async/Filesys: an implementation of fold that follows links
This version of [fold] can follow links while detecting the cycles
they may induce. The traversal should be done with caution, since in
addition to cycles, it may happen that we visit the same object
several times under different names. Here the iterated function will
be called as many times as a link leads to it, but with some
indication that it's been traversed already.

More precisely the iterated function is given an two absolute paths
and a boolean:
  - the first path is the path that follows from the traversal, it may
    contain links, and/or be outside the initial start of the
    traversal (the latter is the reason why we can't provide the
    iterated function a relative path like [fold])
  - the second path is the resolved path of the first one. In
    particular it has no link in it
  - the boolean indicates whether this object has already been visited
    before under a different name.