fold_follows_link: upgrade test, fix test, found bug, fix function
Added a link creating a cycle in the test directory. This showed that
the specification of [fold_follows_link] given in d9b9b4a650 is
wrong. In particular this showed that a given file or directory may
have an infinite number of aliases reachable from a given location. So
it is pointless to try traversing all of them, and that is why the
specification introduced in this commit states that each object should
be visited once. More precisely [fold_follows_link]:

  - traverses a path following links
  - (links being counted as visited objects)
  - visiting each object exactly once
  - providing the alias that happen to be encoutered first and the
    resolved name to the iterated function

The test now passes, but the new configuration of the test directory
makes [fold] loop forever.