async/Filesys: fixed fold for cycles.
[fold] was supposed not to follow links but... it did, because upon
link discovery, it would call itself recursively packing the link path
with a [`File] or a [`Dir] variant. In the [`Dir] case, [fold] would
not check wheter it had a link or not, and ask for the contents of the
directory. Now links are tagged as such, so that [fold] do not recurse
on them.

Maybe it'd be better to have the tag passed to the iterated function?

In any case, the test on fold now passes.